テストで消すファイルがないときのエラーに対応した
```
  1) HealthCheck /custom_route_prefix works with smtp server and valid custom_check
     Failure/Error: FileUtils.rm(CUSTOM_CHECK_FILE_PATH)

     Errno::ENOENT:
       No such file or directory @ unlink_internal - spec/dummy/tmp/custom_file
     # ./spec/spec_helper.rb:35:in `ensure in enable_custom_check'
     # ./spec/spec_helper.rb:35:in `enable_custom_check'
     # ./spec/health_check_spec.rb:6:in `block (3 levels) in <top (required)>'
     # ------------------
     # --- Caused by: ---
     # Errno::ENOENT:
     #   No such file or directory @ rb_sysopen - spec/dummy/tmp/custom_file
     #   ./spec/spec_helper.rb:32:in `write'
```

+++ spec/spec_helper.rb
@@ -32,5 +32,5 @@ def enable_custom_check(&block)
   File.write(CUSTOM_CHECK_FILE_PATH, 'hello')
   block.call
 ensure
-  FileUtils.rm(CUSTOM_CHECK_FILE_PATH)
+  FileUtils.rm(CUSTOM_CHECK_FILE_PATH) if File.exist?(CUSTOM_CHECK_FILE_PATH)
 end
